AREAJAVA-59 Code refactoring
