+++ src/test/java/eu/lundegaard/commons/rest/testclasses/TestFeignClient.java
@@ -1,4 +1,4 @@
-package eu.lundegaard.commons.rest;
+package eu.lundegaard.commons.rest.testclasses;
 
 import feign.RequestLine;
